Logger, better threadpool (maybe)
Implemented a logger that (barely) runs on a separate thread and wanted
to see if another thread pool makes any difference. It definitely is a
better implementation, but I can't say that it has completely floored me

+++ SXIRenderer/src/detail/Scene.cpp
@@ -7,7 +7,7 @@ namespace sxi::renderer::detail
 {
 Scene* scene{};
 
-SceneData::SceneData(u8 maxObjects, u8 frame) : frame(frame)
+SceneData::SceneData(size_t maxObjects, u8 frame) : frame(frame)
 {
     objectDescriptorSets.resize(maxObjects);
     createFrameDescriptorSet();
@@ -93,7 +93,7 @@ void SceneData::createObjectDescriptorSets()
     }
 }
 
-Scene::Scene(u8 maxObjects)
+Scene::Scene(size_t maxObjects)
 {
     objectUBOs.resize(maxObjects);
     for (size_t i = 0; i < sceneDatas.size(); ++i)
